+++ libraries/ndg/ndg_adam.c
@@ -33,6 +33,8 @@
 *        Add some const-ing
 *     4-JUL-2008 (TIMJ):
 *        More const-ing goodness.
+*     15-JUL-2008 (TIMJ):
+*        Use size_t for index to match new Grp interface.
 *     {enter_further_changes_here}
 
 *  Copyright:
@@ -72,7 +74,7 @@
 
 F77_SUBROUTINE(ndg_assoc)( CHARACTER(PARAM), LOGICAL(VERB), INTEGER(IGRP), INTEGER(SIZE), LOGICAL(FLAG), INTEGER(STATUS) TRAIL(PARAM) );
 
-void ndgAssoc( const char * param, int verb, Grp ** igrp, int *size, int * flag, int *status 
+void ndgAssoc( const char * param, int verb, Grp ** igrp, size_t *size, int * flag, int *status 
 ){
    DECLARE_INTEGER(IGRP);
    DECLARE_INTEGER(SIZE);
@@ -104,7 +106,7 @@ void ndgAssoc( const char * param, int verb, Grp ** igrp, int *size, int * flag,
 
 F77_SUBROUTINE(ndg_creat)( CHARACTER(PARAM), INTEGER(IGRP0), INTEGER(IGRP), INTEGER(SIZE), LOGICAL(FLAG), INTEGER(STATUS) TRAIL(PARAM) );
 
-void ndgCreat( const char * param, const Grp *igrp0, Grp ** igrp, int *size, int * flag, int *status 
+void ndgCreat( const char * param, const Grp *igrp0, Grp ** igrp, size_t *size, int * flag, int *status 
 ){
    DECLARE_INTEGER(IGRP0);
    DECLARE_INTEGER(IGRP);
